pyproject.toml
[tool.poetry]
name = "azure-monitor"
version = "0.1.0"
description = ""
authors = ["shin-hama <htysivista@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
azure-devops = "^6.0.0-beta.4"
python-dotenv = "^0.19.2"
SQLAlchemy = "^1.4.29"
PyMySQL = "^1.0.2"
alembic = "^1.7.5"
python-redmine = "^2.3.0"
pytz = "^2021.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
mypy = "^0.930"
black = "^21.12b0"
types-pytz = "^2021.3.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
